+++ DataObjects/Migrations/201506161757460_Initial.cs
@@ -187,7 +187,6 @@ namespace forCrowd.WealthEconomy.DataObjects.Migrations
                         InitialValue = c.Decimal(nullable: false, precision: 18, scale: 2),
                         MainElementId = c.Int(),
                         UseFixedResourcePoolRate = c.Boolean(nullable: false),
-                        EnableSubtotals = c.Boolean(nullable: false),
                         ResourcePoolRate = c.Decimal(precision: 18, scale: 2),
                         ResourcePoolRateCount = c.Int(),
                         CreatedOn = c.DateTime(nullable: false),